+++ R/pits.R
@@ -1,13 +1,12 @@
 #' @title Pit stops of a race or set
 #' @description Pit stop results of a race or set
 #' @param round Number of the race (integer), set of races a:b (vector) o "all" (character)
-#' @param year Year of the race (integer). 2025 or higher
+#' @param year Year of the race (integer). 2019 or higher
 #' @return A tibble containing the pit stops values of the specified race(s)
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' pits(10, 2025)
 #' pits(1:2, 2025)
-#' pits("all",2025)
 #' }
 #' @importFrom f1dataR load_schedule
 #' @importFrom readr read_delim